refactor : client의 login id를 spring context에서 추출하는 메서드 생성
